CommentedOut ToolTips + tests
